requirements.txt
PyYAML~=6.0.2
tweepy==4.14.0